Refactor-Create DateTimeUtil class
